# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: channel
    description: Search and answer from the configured YouTube channel
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /channel/search:
    get:
      operationId: searchChannel
      tags: [channel]
      summary: Analyze the question and search the channel in depth
      description: Analyzes the Arabic question (type, keywords, religious topics), searches the configured YouTube channel, mines available captions with timestamps, and returns a grounded Arabic answer with per-video analysis and a synthesized summary.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 500
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 3
        - name: depth
          in: query
          required: false
          description: How deep the transcript mining goes.
          schema:
            type: string
            enum: [concise, balanced, deep]
            default: balanced
      responses:
        "200":
          description: Grounded channel answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSearchResponse"
        "400":
          description: Invalid question
        "502":
          description: YouTube could not be reached or captions were unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    QuestionAnalysis:
      type: object
      properties:
        kind:
          type: string
        kindLabel:
          type: string
        keywords:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
      required:
        - kind
        - kindLabel
        - keywords
        - topics
    VideoHighlight:
      type: object
      properties:
        text:
          type: string
        time:
          type: string
      required:
        - text
        - time
    ChannelVideo:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        url:
          type: string
        thumbnail:
          type: string
        snippet:
          type: string
        relevance:
          type: number
        transcriptAvailable:
          type: boolean
        topics:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            $ref: "#/components/schemas/VideoHighlight"
      required:
        - videoId
        - title
        - url
        - thumbnail
        - snippet
        - relevance
        - transcriptAvailable
        - topics
        - highlights
    ChannelSearchResponse:
      type: object
      properties:
        answer:
          type: string
        confidence:
          type: number
        matchedTopic:
          type: string
        channelName:
          type: string
        channelUrl:
          type: string
        videos:
          type: array
          items:
            $ref: "#/components/schemas/ChannelVideo"
        sourceStatus:
          type: string
          enum: [live, cached, metadata-only]
        analysis:
          $ref: "#/components/schemas/QuestionAnalysis"
        summaryPoints:
          type: array
          items:
            type: string
        coverage:
          type: number
      required:
        - answer
        - confidence
        - matchedTopic
        - channelName
        - channelUrl
        - videos
        - sourceStatus
        - analysis
        - summaryPoints
        - coverage
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error