# lib/api-zod/src/generated/types/channelSearchResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelSearchResponseSourceStatus } from './channelSearchResponseSourceStatus';
import type { ChannelVideo } from './channelVideo';
import type { QuestionAnalysis } from './questionAnalysis';

export interface ChannelSearchResponse {
  answer: string;
  confidence: number;
  matchedTopic: string;
  channelName: string;
  channelUrl: string;
  videos: ChannelVideo[];
  sourceStatus: ChannelSearchResponseSourceStatus;
  analysis: QuestionAnalysis;
  summaryPoints: string[];
  coverage: number;
}