# lib/api-zod/src/generated/types/questionAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionAnalysis {
  kind: string;
  kindLabel: string;
  keywords: string[];
  topics: string[];
}